Remove reliance on tags from the grammar
The tag faeture of pest has been a little buggy, and so I've ripped it out. This does add a substantial degree of coupling between the exact grammar rules and whether various rules are silenced, but the tags were so often applied to the (to my eyes) wrong pattern that I can't rely on them.

Also cleans up some of the AST generation logic and adds back in a few CommonMark rules, namely that verbatim blocks and headers should interupt paragraphs.